pyproject.toml
[tool.poetry]
name = "test-task-zhushma-artyom"
version = "0.1.0"
description = "Test task"
readme = "README.md"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
python-dotenv = ">=1,<2"
litestar = {extras = ["standard"], version = ">=2,<3"}
litestar-granian = ">=0,<1"
litestar-asyncpg = ">=0,<1"
advanced-alchemy = ">=0.20,<0.21"
msgspec = ">=0.18.6,<0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"
